+++ src/Models/Route.cs
@@ -6,13 +6,13 @@ namespace Models
     {
         [JsonPropertyName("id")] public string Id { get; set; }
         [JsonPropertyName("color")] public string Color { get; set; }
-        [JsonPropertyName("velocity")] public float Velocity { get; set; }
+        [JsonPropertyName("velocity")] public double Velocity { get; set; }
         [JsonPropertyName("path")] public Point[] Path { get; set; }
     }
 
     public class Point
     {
         [JsonPropertyName("busStopIndex")] public int BusStopIndex { get; set; }
-        [JsonPropertyName("duration")] public int Duration { get; set; }
+        [JsonPropertyName("duration")] public double Duration { get; set; }
     }
 }
